+++ src/app/layout.tsx
@@ -1,4 +1,4 @@
-'use client';
+"use client"
 
 import CssBaseline from '@mui/material/CssBaseline';
 import { ThemeProvider } from '@mui/material/styles';
