+++ src/resources/auth/provider.tsx
@@ -1,6 +1,6 @@
 import React, { useReducer, useEffect, ReactNode } from 'react';
 
-import { getAuthToken, setAuthToken, removeAuthToken, setRefreshToken, removeRefreshToken } from '@/lib/utils';
+import { getAuthToken, setAuthToken, removeAuthToken, setRefreshToken, removeRefreshToken } from '@/resources/auth';
 
 import { api, Context, MODELS, DEFAULTS, reducers, isAuthenticated } from '.';
